1. Use "short" as the filter's type. 2. Let bind_io_handle() calls unbind_io_handle() when failures happen. 3. Remove cb from io sessions after filters are unregistered in unbind_io_handle(). 4. Code optimization.

+++ src/core/tools/hpc/io_looper.h
@@ -124,7 +124,7 @@ namespace dsn
             struct epoll_event        _events[IO_LOOPER_MAX_EVENT_COUNT];
 # elif defined(__APPLE__) || defined(__FreeBSD__)
             struct kevent             _events[IO_LOOPER_MAX_EVENT_COUNT];
-            typedef std::unordered_set<int> kqueue_filters;
+            typedef std::unordered_set<short> kqueue_filters;
             kqueue_filters            _filters;
 # endif
